refactor: split fallback.py into package and add 30 unit tests
Break handlers/core/fallback.py (903 lines) into handlers/fallback/ package:
- registration.py, tests.py, training.py, knowledge.py, company.py (domain handlers)
- grouped.py (8 grouped state handlers), universal.py (text + callback fallback)
- Move StateFilter constants to states/filters.py
- Move helper functions to utils/messages/fallback.py
- Add 30 new unit tests covering all previously untested handlers (46 -> 76)
